requirements.txt
streamlit
pandas
matplotlib
seaborn
google-generativeai
reportlab
markdown
python-dotenv
pymupdf